refactor compatibility tests to highlight problems in our plugin
