--- plugin/src/test/fixtures/plugin_groovy/src/main/resources/META-INF/gradle-plugins/com.github.triplet.play.test.properties
@@ -1,1 +1,0 @@
-implementation-class=com.github.triplet.gradle.play.test.AndroidDefaultConfigPlugin
